improving 1D backend
