+++ kymatio/scattering1d/__init__.py
@@ -1,4 +1,3 @@
 __all__ = ['Scattering1D']
 
-
-from kymatio.scattering1d.core.scattering1d import Scattering1D
+from .frontend import Scattering1D
